chore: finish migrating base package to Swan
