+++ swan/src/base/directives.ts
@@ -1,0 +1,30 @@
+/*
+Copyright 2020 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
+/**
+ * This module re-exports commonly used directives from lit.
+ * It serves as a central point for importing lit directives
+ * within the SWC ecosystem.
+ */
+
+export { ifDefined } from 'lit/directives/if-defined.js';
+export { repeat } from 'lit/directives/repeat.js';
+export { classMap } from 'lit/directives/class-map.js';
+export { styleMap } from 'lit/directives/style-map.js';
+export type { StyleInfo } from 'lit/directives/style-map.js';
+export { until } from 'lit/directives/until.js';
+export { live } from 'lit/directives/live.js';
+export { when } from 'lit/directives/when.js';
+export { join } from 'lit/directives/join.js';
+export { unsafeHTML } from 'lit/directives/unsafe-html.js';
+export type { Ref } from 'lit/directives/ref.js';
+export { createRef, ref } from 'lit/directives/ref.js';
