+++ tools/base/src/condition-attribute-with-id.ts
@@ -10,38 +10,14 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-export function conditionAttributeWithoutId(
-    el: HTMLElement,
-    attribute: string,
-    ids: string[]
-): void {
-    const ariaDescribedby = el.getAttribute(attribute);
-    let descriptors = ariaDescribedby ? ariaDescribedby.split(/\s+/) : [];
-    descriptors = descriptors.filter(
-        (descriptor) => !ids.find((id) => descriptor === id)
-    );
-    if (descriptors.length) {
-        el.setAttribute(attribute, descriptors.join(' '));
-    } else {
-        el.removeAttribute(attribute);
-    }
-}
+/**
+ * ⚠️ IMPORTANT: SOURCE OF TRUTH MOVED ⚠️
+ *
+ * The authoritative implementation is now in:
+ * swan/src/base/condition-attribute-with-id.ts
+ *
+ * This file is maintained for backward compatibility ONLY.
+ * DO NOT modify this file directly; instead make changes in Swan.
+ */
 
-export function conditionAttributeWithId(
-    el: HTMLElement,
-    attribute: string,
-    id: string | string[]
-): () => void {
-    const ids = Array.isArray(id) ? id : [id];
-    const ariaDescribedby = el.getAttribute(attribute);
-    const descriptors = ariaDescribedby ? ariaDescribedby.split(/\s+/) : [];
-    const hadIds = ids.every((id) => descriptors.indexOf(id) > -1);
-    if (hadIds)
-        /* c8 ignore next 3 */
-        return (): void => {
-            return;
-        };
-    descriptors.push(...ids);
-    el.setAttribute(attribute, descriptors.join(' '));
-    return () => conditionAttributeWithoutId(el, attribute, ids);
-}
+export * from '@spectrum-web-components/swan/base/condition-attribute-with-id.js';
